# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  handlers.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  models.User:
    type: object
  models.UserResponse:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  services.McpTagRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  services.RegisterMcpServiceRequest:
    properties:
      category:
        type: string
      description:
        type: string
      endpoint:
        type: string
      icon_url:
        type: string
      json_schema:
        type: string
      name:
        type: string
      submitter_id:
        type: integer
      tags:
        items:
          type: integer
        type: array
    required:
    - description
    - endpoint
    - name
    type: object
  services.UpdateUserRoleRequest:
    properties:
      role:
        enum:
        - admin
        - user
        example: admin
        type: string
    required:
    - role
    type: object
  utils.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: success
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 这是一个基于Gin+GORM+SQLite的用户管理系统
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 用户管理系统API
  version: "1.0"
paths:
  /mcp-tags:
    get:
      consumes:
      - application/json
      description: 分页获取 MCP 服务列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: id
        description: 排序字段
        in: query
        name: by
        type: string
      - default: true
        description: 是否升序
        in: query
        name: asc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: '返回分页结果: total, page, list([]models.McpTagResponse)'
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取 MCP 服务列表
      tags:
      - mcp-tags
    post:
      consumes:
      - application/json
      description: 创建 MCP 标签
      parameters:
      - description: 创建 MCP 标签请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.McpTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '返回创建结果: id, name'
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 创建 MCP 标签
      tags:
      - mcp-tags
  /mcp-tags/{id}:
    delete:
      consumes:
      - application/json
      description: 删除 MCP 标签
      parameters:
      - description: MCP 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '返回删除结果: id, name'
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 删除 MCP 标签
      tags:
      - mcp-tags
    put:
      consumes:
      - application/json
      description: 更新 MCP 标签
      parameters:
      - description: MCP 标签ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新 MCP 标签请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.McpTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '返回更新结果: id, name'
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 更新 MCP 标签
      tags:
      - mcp-tags
  /mcps:
    get:
      consumes:
      - application/json
      description: 分页获取 MCP 服务列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: id
        description: 排序字段
        in: query
        name: by
        type: string
      - default: true
        description: 是否升序
        in: query
        name: asc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: '返回分页结果: total, page, list([]models.McpResponse)'
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取 MCP 服务列表
      tags:
      - mcps
    post:
      consumes:
      - application/json
      description: 注册一个新的 MCP 服务，并绑定标签
      parameters:
      - description: MCP 服务信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.RegisterMcpServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回创建成功的 MCP 服务
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 创建 MCP 服务
      tags:
      - mcps
  /mcps/{id}:
    delete:
      description: 根据 ID 删除 MCP 服务
      parameters:
      - description: MCP 服务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 未找到 MCP 服务
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 删除 MCP 服务
      tags:
      - mcps
    get:
      description: 根据 ID 获取 MCP 服务信息（含标签）
      parameters:
      - description: MCP 服务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回 MCP 服务详情
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 未找到 MCP 服务
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取 MCP 服务详情
      tags:
      - mcps
  /users:
    get:
      consumes:
      - application/json
      description: 分页获取用户列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: id
        description: 排序字段
        in: query
        name: by
        type: string
      - default: true
        description: 是否升序
        in: query
        name: asc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: '返回分页结果: total, page, list([]models.UserResponse)'
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 没有权限访问
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取用户列表
      tags:
      - users
    post:
      consumes:
      - application/json
      description: 使用提供的信息创建新用户
      parameters:
      - description: 用户负载
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: '成功: 返回创建的用户数据 (models.UserResponse)'
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 创建用户
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 删除用户
      tags:
      - users
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户详情
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回用户数据 (models.UserResponse)
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 用户未找到
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 通过ID获取用户
      tags:
      - users
    put:
      consumes:
      - application/json
      description: 根据ID更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户负载
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: 返回更新后的用户数据 (models.UserResponse)
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 用户未找到
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 更新用户
      tags:
      - users
  /users/{id}/role:
    put:
      consumes:
      - application/json
      description: 管理员修改用户角色（仅管理员可访问）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回更新后的用户数据 (models.UserResponse)
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 用户未找到
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Bearer: []
      summary: 修改用户角色
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Failed to generate token
          schema:
            $ref: '#/definitions/utils.Response'
      summary: User login
      tags:
      - auth
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"